auth/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
pyjwt>=1.7.0
django-cors-headers>=3.5.0